Add AllowEncodedSlashes to apache
Apache proxy should have 'AllowEncodedSlashes On' to support encoding in v3 rooms. "The AllowEncodedSlashes setting is not inherited by virtual hosts, and virtual hosts are used in many default Apache configurations, such as the one in Ubuntu. The workaround is to add the AllowEncodedSlashes setting inside a <VirtualHost> container (/etc/apache2/sites-available/default in Ubuntu)". Source: https://stackoverflow.com/questions/4390436/need-to-allow-encoded-slashes-on-apache